requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
requests
# LangChain & RAG
langchain
langchain-community
langchain-huggingface
chromadb
sentence-transformers
pypdf
unstructured
